🧹 Extract Pixel class from PixelCanvas.tsx
- Moved `Pixel` class to `src/components/effects/PixelCanvas/Pixel.ts`
- Updated `PixelCanvas.tsx` to import `Pixel`
- Added unit tests in `src/components/effects/PixelCanvas/Pixel.test.ts`
- Verified logic with standalone script
