+++ src/components/effects/PixelCanvas/PixelCanvas.tsx
@@ -1,124 +1,6 @@
 import { useEffect, useMemo, useRef } from "react";
 import { debounce } from "../../../utils/commonUtils";
-
-class Pixel {
-  width: number;
-  height: number;
-  ctx: CanvasRenderingContext2D;
-  x: number;
-  y: number;
-  color: string;
-  speed: number;
-  size: number;
-  sizeStep: number;
-  minSize: number;
-  maxSizeInteger: number;
-  maxSize: number;
-  alpha: number;
-  delay: number;
-  counter: number;
-  counterStep: number;
-  isIdle: boolean;
-  isReverse: boolean;
-  isShimmer: boolean;
-
-  constructor(
-    canvas: HTMLCanvasElement,
-    context: CanvasRenderingContext2D,
-    x: number,
-    y: number,
-    color: string,
-    speed: number,
-    delay: number,
-  ) {
-    this.width = canvas.width;
-    this.height = canvas.height;
-    this.ctx = context;
-    this.x = x;
-    this.y = y;
-    this.color = color;
-    this.speed = Pixel.getRandomValue(0.05, 0.6) * speed;
-    this.size = 0;
-    this.sizeStep = Math.random() * 0.2;
-    this.minSize = 0.3;
-    this.maxSizeInteger = 1.4;
-    this.maxSize = Pixel.getRandomValue(this.minSize, this.maxSizeInteger);
-    this.alpha = Pixel.getRandomValue(0.08, 0.25);
-    this.delay = delay;
-    this.counter = 0;
-    this.counterStep = Math.random() * 4 + (this.width + this.height) * 0.01;
-    this.isIdle = false;
-    this.isReverse = false;
-    this.isShimmer = false;
-  }
-
-  static getRandomValue(min: number, max: number) {
-    return Math.random() * (max - min) + min;
-  }
-
-  draw() {
-    const centerOffset = this.maxSizeInteger * 0.5 - this.size * 0.5;
-
-    // Performance: Removed ctx.save()/restore() to reduce draw call overhead
-    // Safe because we explicitly set state properties (alpha, color) for every pixel
-    this.ctx.globalAlpha = this.alpha;
-    this.ctx.fillStyle = this.color;
-    this.ctx.fillRect(
-      this.x + centerOffset,
-      this.y + centerOffset,
-      this.size,
-      this.size,
-    );
-  }
-
-  appear() {
-    this.isIdle = false;
-
-    if (this.counter <= this.delay) {
-      this.counter += this.counterStep;
-      return;
-    }
-
-    if (this.size >= this.maxSize) {
-      this.isShimmer = true;
-    }
-
-    if (this.isShimmer) {
-      this.shimmer();
-    } else {
-      this.size += this.sizeStep;
-    }
-
-    this.draw();
-  }
-
-  disappear() {
-    this.isShimmer = false;
-    this.counter = 0;
-
-    if (this.size <= 0) {
-      this.isIdle = true;
-      return;
-    }
-
-    this.size -= 0.06;
-    this.draw();
-  }
-
-  shimmer() {
-    if (this.size >= this.maxSize) {
-      this.isReverse = true;
-    } else if (this.size <= this.minSize) {
-      this.isReverse = false;
-    }
-
-    if (this.isReverse) {
-      this.size -= this.speed;
-    } else {
-      this.size += this.speed;
-    }
-  }
-}
+import { Pixel } from "./Pixel";
 
 const clampGap = (value: number) => {
   const min = 4;
